// src/app/shared/models/diluyente.model.ts
export class DetalleDiluyente {
    desAplicacionCada: string;
    desCortaDiluyente: string;
    desTiempoInfusion: string;
    desViaAdministracion: string;
    fechaAplicacion: string;
    idMezcla: number;
    idMezclaMedicDiluy: number;
    numTotalDosis: number;
    numDosisDiluyente: number;
    refUnidadMinMedida: string;
    refVelInfusion: string;
    
    totalDosis?: number;

}